config/requirements.txt
Django==4.1.7
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2==2.9.5
djangorestframework-simplejwt==5.2.2
